+++ core/ajax/Diagral_eOne.ajax.php
@@ -25,14 +25,26 @@ try {
     }
 
     ajax::init();
+    // Lancement de la synchronisation des equipements
     if (init('action') == 'synchronize') {
       try {
 		    Diagral_eOne::synchronize();
 		    ajax::success();
       } catch (Exception $e) {
-        ajax::error(displayExeption($e), $e->getCode());
+            ajax::error(displayExeption($e), $e->getCode());
       }
     }
+
+    //Lancement de la suppression des données de tracking
+    if (init('action') == 'delete_remote_datainfo') {
+        try {
+            Diagral_eOne::installTracking(1);
+            ajax::success();
+        } catch (Exception $e) {
+            ajax::error(displayExeption($e), $e->getCode());
+        }
+    }
+
 
     if (init('action') == 'postSave') {
         //Called after a plugin configuration save
@@ -47,6 +59,9 @@ try {
             config::save('retry', init('retry'), 'Diagral_eOne');
             config::save('waitRetry', init('waitRetry'), 'Diagral_eOne');
             config::save('polling_interval', init('polling_interval'), 'Diagral_eOne');
+            config::save('InstallBaseStatus', init('InstallBaseStatus'), 'Diagral_eOne');
+            config::save('InstallBaseAnonymousOnly', init('InstallBaseAnonymousOnly'), 'Diagral_eOne');
+            config::save('InstallBaseEmailAddr', init('InstallBaseEmailAddr'), 'Diagral_eOne');
     		//Let's then the error details
     		ajax::error(displayExeption($e), $e->getCode());
         }
@@ -82,6 +97,127 @@ try {
         }
         ajax::success();
     }
+
+    //Lancement de la suppression des données de tracking
+    if (init('action') == 'notificationVerifyScenario') {
+        try {
+            $alarmEq = eqLogic::byId(init('eqID'));
+            $scenarioID = $alarmEq->getConfiguration('notificationScenarioID');
+            $scenario = scenario::byId($scenarioID);
+            if (is_object($scenario)) {
+                $return = array('scenarioExist' => true, 'scenarioID' => $scenarioID, 'scenarioName' => $scenario->getName());
+                ajax::success(json_encode($return));
+            } else {
+                ajax::success(false);
+            }
+
+        } catch (Exception $e) {
+            ajax::error(displayExeption($e), $e->getCode());
+        }
+    }
+
+    //Création ou mise à jour du scénario de notification
+    if (init('action') == 'notificationGenerateUpdateScenario') {
+        try {
+                $alarmEq = eqLogic::byId(init('eqID'));
+                $notificationPlugin = $alarmEq->getConfiguration('notificationPlugin');
+
+                if ($notificationPlugin === "" || $alarmEq->getConfiguration('notificationEqLogic') === "") {
+                    ajax::error("Le plugin ou l'equipement de notification n'est pas disponible. Pensez à sauvegarder vos configurations avant la création du scénario");
+                }
+
+                //Ouverture du fichier de template
+                $scenarioTemplateLocation = '/../config/scenarioTemplate.json';
+                if (!file_exists(dirname(__FILE__) . $scenarioTemplateLocation)) {
+                    log::add('Diagral_eOne', 'warning', 'Unable to read scenarioTemplate.json');
+                }
+                $scenarioTemplate = json_decode(file_get_contents(dirname(__FILE__) . $scenarioTemplateLocation), true);
+                log::add('Diagral_eOne', 'debug', 'Data : '.var_export($scenarioTemplate, True));
+                if (!is_array($scenarioTemplate)) {
+                    log::add('Diagral_eOne', 'warning', 'Unable to decode plugin file scenarioTemplate.json');
+                }
+
+                // Debug des données reçus
+                log::add('Diagral_eOne', 'debug', 'ID de l\'équipement reçu : ' . init('eqID'));
+
+                // Définission des commandes par plugins supportés
+                switch ($notificationPlugin) {
+                    case 'maillistener':
+                        $trigger = "[HTML]";
+                        $mailTo = "[Expéditeur]";
+                        $subject = "[Sujet]";
+                        $htmlContent = $trigger;
+                        break;
+                    default:
+                        log::add('Diagral_eOne', 'warning', 'Plugin ' . $notificationPlugin . ' not supported to create or update scenario');
+                        ajax::error('Diagral_eOne', 'warning', 'Plugin ' . $notificationPlugin . ' not supported to create or update scenario');
+                        break;
+                }
+
+                // Récupération du nom de la commande de notication (recu dans un format type #eqLogic61#)
+                preg_match('/#eqLogic([0-9]*)#/', $alarmEq->getConfiguration('notificationEqLogic'), $notificationEqID);
+                $notificationEqLogic = eqLogic::byId($notificationEqID[1]);
+                log::add('Diagral_eOne', 'debug', 'Commande : ' . $notificationEqLogic->getHumanName());
+                // Recuperation de l'équipement alarme
+                $alarmEq = eqLogic::byId(init('eqID'));
+
+                // Remplacement des commandes dans le template
+                $scenarioTemplate['trigger'] = array("#" . $notificationEqLogic->getHumanName().$trigger . "#");
+                $scenarioTemplate['elements'][0]['subElements'][0]['expressions'][0]['expression'] = str_replace("#MAIL-TO#", "#" . $notificationEqLogic->getHumanName().$mailTo . "#", $scenarioTemplate['elements'][0]['subElements'][0]['expressions'][0]['expression']);
+                $scenarioTemplate['elements'][0]['subElements'][1]['expressions'][0]['expression'] = "#" . $alarmEq->getHumanName() . "[Importer Message]#";
+                $scenarioTemplate['elements'][0]['subElements'][1]['expressions'][0]['options']['title'] = str_replace("#MAIL-SUBJECT#", "#" . $notificationEqLogic->getHumanName().$subject . "#", $scenarioTemplate['elements'][0]['subElements'][1]['expressions'][0]['options']['title']);
+                $scenarioTemplate['elements'][0]['subElements'][1]['expressions'][0]['options']['message'] = "#" . $notificationEqLogic->getHumanName().$htmlContent . "#";
+
+                // Creation du scenario (ou mise à jour si existant)
+                $scenario = scenario::byId($alarmEq->getConfiguration('notificationScenarioID'));
+                if (!is_object($scenario)) {
+                    // Déclaration d'un nouveau scenario
+                    $scenario = new scenario();
+                }
+                utils::a2o($scenario, $scenarioTemplate);
+                $scenario->setConfiguration('timeDependency', $scenarioTemplate['configuration']['timeDependency']);
+                $scenario->setConfiguration('has_return', $scenarioTemplate['configuration']['has_return']);
+                $scenario_element_list = array();
+                if (isset($scenarioTemplate['elements'])) { // Ajout des elements du scénario
+                    foreach ($scenarioTemplate['elements'] as $element) {
+                        $scenario_element_list[] = scenarioElement::saveAjaxElement($element);
+                    }
+                    $scenario->setScenarioElement($scenario_element_list);
+                }
+                $scenario->save();
+
+                // Retrouve le scénario venant d'être créé pour stocker son ID
+                $scenarioCreated = scenario::byObjectNameGroupNameScenarioName('Aucun', 'Aucun', 'Notification Diagral');
+                $alarmEq->setConfiguration('notificationScenarioID', $scenarioCreated->getId());
+                $alarmEq->save();
+
+                log::add('Diagral_eOne', 'debug', 'Data : '.var_export($scenarioCreated, True));
+                ajax::success();
+
+        } catch (Exception $e) {
+            ajax::error(displayExeption($e), $e->getCode());
+        }
+    }
+
+    // Suppression du scénario de recéption de notification Diagral
+    if (init('action') == 'notificationDeleteScenario') {
+        try {
+            // Recuperation de l'équipement alarme
+            $alarmEq = eqLogic::byId(init('EqId'));
+            // Recuperation de l'ID du plugin Alarm
+            $scenario = scenario::byId($alarmEq->getConfiguration('notificationScenarioID'));
+            log::add('Diagral_eOne', 'debug', 'Data : '.var_export($scenario, True));
+            if (is_object($scenario)) {
+                $scenario->remove();
+                $alarmEq->setConfiguration('notificationScenarioID', '');
+            } else {
+                ajax::error("Scenario invalide");
+            }
+            ajax::success();
+        } catch (Exception $e) {
+            ajax::error(displayExeption($e), $e->getCode());
+        }
+      }
 
 
     throw new Exception(__('Aucune méthode correspondante à : ', __FILE__) . init('action'));